test : 좌석예약 동시성 테스트
